requirements.txt
Django>=4.2
psycopg2-binary>=2.9
groq>=0.11.0
pdfminer.six>=20221105
python-pptx>=0.6.21
python-dotenv>=1.0.0
gunicorn>=20.1.0
requests>=2.28.0
dj-database-url>=1.0.0
